refactor(app2): rename schemas -> schema
